Create README.md
Create a 3D illusion for a profile picture of a٫٫ پسر یا دختر ٫٫ in a casual black shirt on a back seat with a black burgundy with smoke and the flag of Afghanistan, with the flag of Afghanistan logo and a crown in a vest effect flag. He looks forward in sneakers, a black cricket cap, a Rolex watch, and sunglasses. On the blouse it shows "Afghanistan" in white font. It should not have a shadow and it has wings as if it were an angel.